Tested all Methods of controllerLogic. Add a main class to test future problems and fixing a problem of Script
